renamed package from articlePage to articleListPage
