+++ src/main/java/org/viesure/articleListPage/ArticleListPageLocators.java
@@ -1,4 +1,4 @@
-package org.viesure.articlePage;
+package org.viesure.articleListPage;
 
 import io.appium.java_client.android.AndroidDriver;
 import io.appium.java_client.android.AndroidElement;
